Compile tests on Scala Native (#6283)
* Scala Native 0.4.3

* Minimize changes

* Use scala-java-time instead of sjavatime for Scala Native

* Run tests on Scala Native, not just compile

* Update plugins.sbt

* native-loop-core 0.2.1

* Update ci.yml

* format

* fix Scala 2.11

* fix Scala 2.11

* fix Scala 2.11

* More memory for non-JVM processes

* Just build tests, don't run them

* Add nativeSettings to streamsTests and testTests

* fix 2.11 Native

* fix 2.11 Native

* -Xmx4G
