+++ .github/workflows/ci.yml
@@ -1,8 +1,8 @@
 name: CI
 
 env:
-  JDK_JAVA_OPTIONS: -XX:+PrintCommandLineFlags -Xmx3G # JDK_JAVA_OPTIONS is _the_ env. variable to use for modern Java
-  JVM_OPTS: -XX:+PrintCommandLineFlags -Xmx3G # for Java 8 only (sadly, it is not modern enough for JDK_JAVA_OPTIONS)
+  JDK_JAVA_OPTIONS: -XX:+PrintCommandLineFlags -Xmx4G # JDK_JAVA_OPTIONS is _the_ env. variable to use for modern Java
+  JVM_OPTS: -XX:+PrintCommandLineFlags -Xmx4G # for Java 8 only (sadly, it is not modern enough for JDK_JAVA_OPTIONS)
 
 on:
   pull_request:
@@ -27,6 +27,8 @@ jobs:
         check-latest: true
     - name: Cache scala dependencies
       uses: coursier/cache-action@v6
+    - name: Install libuv
+      run: sudo apt-get update && sudo apt-get install -y libuv1-dev
     - name: Lint code
       run: ./sbt check
 
@@ -44,6 +46,8 @@ jobs:
           check-latest: true
       - name: Cache scala dependencies
         uses: coursier/cache-action@v6
+      - name: Install libuv
+        run: sudo apt-get update && sudo apt-get install -y libuv1-dev
       - name: Check generation of ScalaDoc
         run: ./sbt +Test/compile
 
@@ -61,6 +65,8 @@ jobs:
           check-latest: true
       - name: Cache scala dependencies
         uses: coursier/cache-action@v6
+      - name: Install libuv
+        run: sudo apt-get update && sudo apt-get install -y libuv1-dev
       - name: Check that building packages works
         run: ./sbt +publishLocal
 
@@ -78,6 +84,8 @@ jobs:
           check-latest: true
       - name: Cache scala dependencies
         uses: coursier/cache-action@v6
+      - name: Install libuv
+        run: sudo apt-get update && sudo apt-get install -y libuv1-dev
       - name: Check Website Generation
         run: ./sbt docs/docusaurusCreateSite
         
@@ -153,6 +161,8 @@ jobs:
           check-latest: true
       - name: Cache scala dependencies
         uses: coursier/cache-action@v6
+      - name: Install libuv
+        run: sudo apt-get update && sudo apt-get install -y libuv1-dev
       - name: Test on different Scala target platforms
         run: ./sbt test${{ matrix.platform }}
 
@@ -179,6 +189,8 @@ jobs:
           distribution: temurin
           java-version: 8
           check-latest: true
+      - name: Install libuv
+        run: sudo apt-get update && sudo apt-get install -y libuv1-dev
       - name: Release
         run: ./sbt ci-release
         env:
