+++ build.sbt
@@ -36,7 +36,7 @@ addCommandAlias(
 )
 addCommandAlias(
   "testNative",
-  ";coreNative/test;stacktracerNative/test;streamsNative/test;testNative/test;testRunnerNative/test;concurrentNative/test" // `test` currently executes only compilation, see `nativeSettings` in `BuildHelper`
+  ";coreTestsNative/test;stacktracerNative/test;streamsTestsNative/test;testTestsNative/test;concurrentNative/test" // `test` currently executes only compilation, see `nativeSettings` in `BuildHelper`
 )
 addCommandAlias(
   "testJVM",
@@ -93,6 +93,7 @@ lazy val root = project
     coreNative,
     coreTestsJS,
     coreTestsJVM,
+    coreTestsNative,
     docs,
     examplesJS,
     examplesJVM,
@@ -108,6 +109,7 @@ lazy val root = project
     streamsNative,
     streamsTestsJS,
     streamsTestsJVM,
+    streamsTestsNative,
     testJS,
     testJVM,
     testNative,
@@ -126,7 +128,8 @@ lazy val root = project
     testScalaCheckJVM,
     testScalaCheckNative,
     testTestsJS,
-    testTestsJVM
+    testTestsJVM,
+    testTestsNative
   )
   .enablePlugins(ScalaJSPlugin)
 
@@ -164,7 +167,7 @@ lazy val coreNative = core.native
     )
   )
 
-lazy val coreTests = crossProject(JSPlatform, JVMPlatform)
+lazy val coreTests = crossProject(JSPlatform, JVMPlatform, NativePlatform)
   .in(file("core-tests"))
   .dependsOn(core)
   .dependsOn(test)
@@ -193,6 +196,9 @@ lazy val coreTestsJS = coreTests.js
       }
     }
   )
+
+lazy val coreTestsNative = coreTests.native
+  .settings(nativeSettings)
 
 lazy val macros = crossProject(JSPlatform, JVMPlatform)
   .in(file("macros"))
@@ -239,7 +245,7 @@ lazy val streamsJS = streams.js
 lazy val streamsNative = streams.native
   .settings(nativeSettings)
 
-lazy val streamsTests = crossProject(JSPlatform, JVMPlatform)
+lazy val streamsTests = crossProject(JSPlatform, JVMPlatform, NativePlatform)
   .in(file("streams-tests"))
   .dependsOn(streams)
   .dependsOn(coreTests % "test->test;compile->compile")
@@ -267,6 +273,9 @@ lazy val streamsTestsJS = streamsTests.js
       }
     }
   )
+
+lazy val streamsTestsNative = streamsTests.native
+  .settings(nativeSettings)
 
 lazy val test = crossProject(JSPlatform, JVMPlatform, NativePlatform)
   .in(file("test"))
@@ -301,7 +310,7 @@ lazy val testNative = test.native
     )
   )
 
-lazy val testTests = crossProject(JSPlatform, JVMPlatform)
+lazy val testTests = crossProject(JSPlatform, JVMPlatform, NativePlatform)
   .in(file("test-tests"))
   .dependsOn(test)
   .settings(stdSettings("test-tests"))
@@ -320,6 +329,8 @@ lazy val testTestsJS = testTests.js
       ("org.scala-js" %%% "scalajs-java-securerandom" % "1.0.0").cross(CrossVersion.for3Use2_13)
     )
   )
+lazy val testTestsNative = testTests.native
+  .settings(nativeSettings)
 
 lazy val testMagnolia = crossProject(JVMPlatform, JSPlatform)
   .in(file("test-magnolia"))
@@ -398,9 +409,10 @@ lazy val testScalaCheck = crossProject(JSPlatform, JVMPlatform, NativePlatform)
     )
   )
 
-lazy val testScalaCheckJVM    = testScalaCheck.jvm
-lazy val testScalaCheckJS     = testScalaCheck.js
-lazy val testScalaCheckNative = testScalaCheck.native.settings(nativeSettings)
+lazy val testScalaCheckJVM = testScalaCheck.jvm
+lazy val testScalaCheckJS  = testScalaCheck.js
+lazy val testScalaCheckNative = testScalaCheck.native
+  .settings(nativeSettings)
 
 lazy val stacktracer = crossProject(JSPlatform, JVMPlatform, NativePlatform)
   .in(file("stacktracer"))
