+++ test-tests/shared/src/test/scala/zio/test/CompileSpec.scala
@@ -12,14 +12,6 @@ object CompileSpec extends ZIOBaseSpec {
       val expected = "value ++ is not a member of Int"
       if (TestVersion.isScala2) assertM(typeCheck("1 ++ 1"))(isLeft(equalTo(expected)))
       else assertM(typeCheck("1 ++ 1"))(isLeft(anything))
-    },
-    testM("typeCheck expansion must use Left and Right from scala.util") {
-      // if this does not compile the macro expansion probably picks up the wrong Left or Right
-      case class Left(s: Nothing)
-      case class Right(s: Nothing)
-      val assertRight = assertM(typeCheck("1 + 1"))(isRight(anything))
-      val assertLeft  = assertM(typeCheck("1 ++ 1"))(isLeft(anything))
-      (assertLeft <*> assertRight).map { case (l, r) => l && r }
     }
   )
 }
